Organização e ajuste do código. O exercício está na segunda master.

+++ src/main/java/com/exampleRelacaoEntreEstadoECidades/relacaoentreestadosecidades/repository/EstadosRepository.java
@@ -3,5 +3,5 @@ package com.exampleRelacaoEntreEstadoECidades.relacaoentreestadosecidades.reposi
 import com.exampleRelacaoEntreEstadoECidades.relacaoentreestadosecidades.model.EstadosModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstadosRepository extends JpaRepository<EstadosModel, Long> {
+public interface EstadosRepository extends JpaRepository<EstadosModel, Integer> {
 }
